+++ Makefile
@@ -1,2 +1,5 @@
 running:
 	CompileDaemon -build="go build -o ./cmd/main ./cmd" -command=./cmd/main
+
+run:
+	go run ./cmd/main.go
